requirements.txt
opencv-python
opencv-contrib-python
Pillow
numpy
tkcalendar
mysql-connector-python
